simulations/tier-a-comparator/pyproject.toml
[project]
name = "tier-a-comparator"
version = "0.1.0"
description = "CERES Tier A scenario comparator"
requires-python = ">=3.12"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=9.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.setuptools.packages.find]
where = ["src"]
